Minor corrections to the logging and exception throwing macroses
Logging and exception throwing macroses can now take both pointers or
references of the context object

+++ lexgine/core/exception.cpp
@@ -22,7 +22,7 @@ std::string const& Exception::throwingFunctionName() const
     return m_function_name;
 }
 
-uint32_t Exception::lineNumber() const
+long Exception::lineNumber() const
 {
     return m_line_number;
 }
